requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
apscheduler==3.10.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==42.0.0
wechatpy==1.8.18
xmltodict==0.13.0
gunicorn==21.2.0
build==1.0.3
pyjwt==2.8.0
